// Angular/src/app/state/actions/action.types.enum.ts
export enum ActionTypes {
  LOAD_CONTACTS_REQUEST = '[Contact] Load Contacts Request',
  LOAD_CONTACTS_FAILURE = '[Contact] Load Contacts Failure',
  LOAD_CONTACTS_SUCCESS = '[Contact] Load Contacts Success',
  ADD_CONTACT_REQUEST = '[Contact] Add Contact Request',
  ADD_CONTACT_FAILURE = '[Contact] Add Contact Failure',
  ADD_CONTACT_SUCCESS = '[Contact] Add Contact Success',
  UPDATE_CONTACT_REQUEST = '[Contact] Update Contact Request',
  UPDATE_CONTACT_FAILURE = '[Contact] Update Contact Failure',
  UPDATE_CONTACT_SUCCESS = '[Contact] Update Contact Success',
  REMOVE_CONTACT_REQUEST = '[Contact] Remove Contact Request',
  REMOVE_CONTACT_FAILURE = '[Contact] Remove Contact Failure',
  REMOVE_CONTACT_SUCCESS = '[Contact] Remove Contact Success',
}